Conflicts:
	src/EventStore/EventStore.Core/Services/Monitoring/SystemStatsHelper.cs
	src/EventStore/EventStore.Core/Services/UserManagement/UserManagementService.cs
	src/EventStore/EventStore.Projections.Core.Tests/Services/core_projection/projection_checkpoint/when_requesting_checkpoint_after_all_writes_completed.cs
	src/EventStore/EventStore.Projections.Core/Services/Processing/EmittedStream.cs
	src/EventStore/EventStore.sln

+++ src/EventStore/EventStore.Core/Services/Storage/StorageWriterService.cs
@@ -28,7 +28,9 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.IO;
 using System.Linq;
+using System.Text;
 using System.Threading;
 using EventStore.Common.Log;
 using EventStore.Common.Utils;
@@ -41,6 +43,8 @@ using EventStore.Core.Services.Storage.EpochManager;
 using EventStore.Core.Services.Storage.ReaderIndex;
 using EventStore.Core.TransactionLog.Chunks;
 using EventStore.Core.TransactionLog.LogRecords;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 
 namespace EventStore.Core.Services.Storage
 {
@@ -63,7 +67,7 @@ namespace EventStore.Core.Services.Storage
 
         protected readonly TFChunkDb Db;
         protected readonly TFChunkWriter Writer;
-        protected readonly IIndexWriter IndexWriter;
+        private readonly IIndexWriter _indexWriter;
         protected readonly IEpochManager EpochManager;
 
         protected readonly IPublisher Bus;
@@ -72,7 +76,7 @@ namespace EventStore.Core.Services.Storage
         private readonly InMemoryBus _writerBus;
 
         private readonly Stopwatch _watch = Stopwatch.StartNew();
-        private readonly int _minFlushDelay;
+        private readonly double _minFlushDelay;
         private long _lastFlushDelay;
         private long _lastFlushTimestamp;
 
@@ -93,7 +97,7 @@ namespace EventStore.Core.Services.Storage
 
         public StorageWriterService(IPublisher bus, 
                                     ISubscriber subscribeToBus,
-                                    int minFlushDelayMs,
+                                    TimeSpan minFlushDelay,
                                     TFChunkDb db,
                                     TFChunkWriter writer, 
                                     IIndexWriter indexWriter,
@@ -101,7 +105,6 @@ namespace EventStore.Core.Services.Storage
         {
             Ensure.NotNull(bus, "bus");
             Ensure.NotNull(subscribeToBus, "subscribeToBus");
-            Ensure.Nonnegative(minFlushDelayMs, "minFlushDelayMs");
             Ensure.NotNull(db, "db");
             Ensure.NotNull(writer, "writer");
             Ensure.NotNull(indexWriter, "indexWriter");
@@ -110,10 +113,10 @@ namespace EventStore.Core.Services.Storage
             Bus = bus;
             _subscribeToBus = subscribeToBus;
             Db = db;
-            IndexWriter = indexWriter;
+            _indexWriter = indexWriter;
             EpochManager = epochManager;
 
-            _minFlushDelay = minFlushDelayMs * TicksPerMs;
+            _minFlushDelay = minFlushDelay.TotalMilliseconds * TicksPerMs;
             _lastFlushDelay = 0;
             _lastFlushTimestamp = _watch.ElapsedTicks;
 
@@ -156,7 +159,7 @@ namespace EventStore.Core.Services.Storage
             {
                 StorageWriterQueue.Stop();
                 BlockWriter = true;
-                Bus.Publish(new SystemMessage.ServiceShutdown("StorageWriterService"));
+                Bus.Publish(new SystemMessage.ServiceShutdown("StorageWriter"));
             }
         }
 
@@ -190,13 +193,7 @@ namespace EventStore.Core.Services.Storage
 
         void IHandle<SystemMessage.SystemInit>.Handle(SystemMessage.SystemInit message)
         {
-            // We rebuild index till the chaser position, because
-            // everything else will be done by chaser as during replication
-            // with no concurrency issues with writer, as writer before jumping 
-            // into master-mode and accepting writes will wait till chaser caught up.
-            var chaserCheckpoint = Db.Config.ChaserCheckpoint.Read();
-            IndexWriter.Init(chaserCheckpoint);
-            Bus.Publish(new SystemMessage.StorageWriterInitializationDone());   
+            Bus.Publish(new SystemMessage.ServiceInitialized("StorageWriter"));
         }
 
         public virtual void Handle(SystemMessage.StateChangeMessage message)
@@ -207,7 +204,7 @@ namespace EventStore.Core.Services.Storage
             {
                 case VNodeState.Master:
                 {
-                    IndexWriter.Reset();
+                    _indexWriter.Reset();
                     EpochManager.WriteNewEpoch(); // forces flush
                     break;
                 }
@@ -254,33 +251,35 @@ namespace EventStore.Core.Services.Storage
             Bus.Publish(new SystemMessage.WaitForChaserToCatchUp(message.CorrelationId, totalTime));
         }
 
-        void IHandle<StorageMessage.WritePrepares>.Handle(StorageMessage.WritePrepares message)
+        void IHandle<StorageMessage.WritePrepares>.Handle(StorageMessage.WritePrepares msg)
         {
             Interlocked.Decrement(ref FlushMessagesInQueue);
 
             try
             {
-                if (message.LiveUntil < DateTime.UtcNow)
+                if (msg.LiveUntil < DateTime.UtcNow)
                     return;
 
-                var commitCheck = IndexWriter.CheckCommit(message.EventStreamId, message.ExpectedVersion, message.Events.Select(x => x.EventId));
+                string streamId = msg.EventStreamId;
+                var commitCheck = _indexWriter.CheckCommit(streamId, msg.ExpectedVersion, msg.Events.Select(x => x.EventId));
                 if (commitCheck.Decision != CommitDecision.Ok)
                 {
-                    ActOnCommitCheckFailure(message.Envelope, message.CorrelationId, commitCheck);
+                    ActOnCommitCheckFailure(msg.Envelope, msg.CorrelationId, commitCheck);
                     return;
                 }
+
                 var prepares = new List<PrepareLogRecord>();
                 var logPosition = Writer.Checkpoint.ReadNonFlushed();
-                if (message.Events.Length > 0)
+                if (msg.Events.Length > 0)
                 {
                     var transactionPosition = logPosition;
-                    for (int i = 0; i < message.Events.Length; ++i)
+                    for (int i = 0; i < msg.Events.Length; ++i)
                     {
-                        var evnt = message.Events[i];
+                        var evnt = msg.Events[i];
                         var flags = PrepareFlags.Data | PrepareFlags.IsCommitted;
                         if (i == 0)
                             flags |= PrepareFlags.TransactionBegin;
-                        if (i == message.Events.Length - 1)
+                        if (i == msg.Events.Length - 1)
                             flags |= PrepareFlags.TransactionEnd;
                         if (evnt.IsJson)
                             flags |= PrepareFlags.IsJson;
@@ -288,8 +287,8 @@ namespace EventStore.Core.Services.Storage
                         // when IsCommitted ExpectedVersion is always explicit
                         var expectedVersion = commitCheck.CurrentVersion + i;
                         var res = WritePrepareWithRetry(
-                            LogRecord.Prepare(logPosition, message.CorrelationId, evnt.EventId,
-                                              transactionPosition, i, message.EventStreamId,
+                            LogRecord.Prepare(logPosition, msg.CorrelationId, evnt.EventId,
+                                              transactionPosition, i, streamId,
                                               expectedVersion, flags, evnt.EventType, evnt.Data, evnt.Metadata));
                         logPosition = res.NewPos;
                         if (i == 0)
@@ -299,14 +298,22 @@ namespace EventStore.Core.Services.Storage
                 }
                 else
                 {
-                    var res = WritePrepareWithRetry(
-                        LogRecord.Prepare(logPosition, message.CorrelationId, Guid.NewGuid(), logPosition, -1,
-                                          message.EventStreamId, message.ExpectedVersion,
+                    WritePrepareWithRetry(
+                        LogRecord.Prepare(logPosition, msg.CorrelationId, Guid.NewGuid(), logPosition, -1,
+                                          streamId, commitCheck.CurrentVersion,
                                           PrepareFlags.TransactionBegin | PrepareFlags.TransactionEnd | PrepareFlags.IsCommitted,
                                           null, Empty.ByteArray, Empty.ByteArray));
-                    prepares.Add(res.Prepare);
                 }
-                IndexWriter.PreCommit(prepares);
+
+                bool softUndeleteMetastream = SystemStreams.IsMetastream(streamId)
+                                              && _indexWriter.IsSoftDeleted(SystemStreams.OriginalStreamOf(streamId));
+
+                _indexWriter.PreCommit(prepares);
+
+                if (commitCheck.IsSoftDeleted)
+                    SoftUndeleteStream(streamId, commitCheck.CurrentVersion + 1);
+                if (softUndeleteMetastream)
+                    SoftUndeleteMetastream(streamId);
             }
             catch (Exception exc)
             {
@@ -319,6 +326,61 @@ namespace EventStore.Core.Services.Storage
             }
         }
 
+
+        private void SoftUndeleteMetastream(string metastreamId)
+        {
+            var origStreamId = SystemStreams.OriginalStreamOf(metastreamId);
+            var rawMetaInfo = _indexWriter.GetStreamRawMeta(origStreamId);
+            SoftUndeleteStream(origStreamId, rawMetaInfo.MetaLastEventNumber, rawMetaInfo.RawMeta,
+                               recreateFrom: _indexWriter.GetStreamLastEventNumber(origStreamId) + 1);
+        }
+
+        private void SoftUndeleteStream(string streamId, int recreateFromEventNumber)
+        {
+            var rawInfo = _indexWriter.GetStreamRawMeta(streamId);
+            SoftUndeleteStream(streamId, rawInfo.MetaLastEventNumber, rawInfo.RawMeta, recreateFromEventNumber);
+        }
+
+        private void SoftUndeleteStream(string streamId, int metaLastEventNumber, byte[] rawMeta, int recreateFrom)
+        {
+            byte[] modifiedMeta;
+            if (!SoftUndeleteRawMeta(rawMeta, recreateFrom, out modifiedMeta))
+                return;
+
+            var logPosition = Writer.Checkpoint.ReadNonFlushed();
+            var res = WritePrepareWithRetry(
+                LogRecord.Prepare(logPosition, Guid.NewGuid(), Guid.NewGuid(), logPosition, 0,
+                                  SystemStreams.MetastreamOf(streamId), metaLastEventNumber,
+                                  PrepareFlags.SingleWrite | PrepareFlags.IsCommitted | PrepareFlags.IsJson,
+                                  SystemEventTypes.StreamMetadata, modifiedMeta, Empty.ByteArray));
+
+            _indexWriter.PreCommit(new[] { res.Prepare });
+        }
+
+        public bool SoftUndeleteRawMeta(byte[] rawMeta, int recreateFromEventNumber, out byte[] modifiedMeta)
+        {
+            try
+            {
+                var jobj = JObject.Parse(Encoding.UTF8.GetString(rawMeta));
+                jobj[SystemMetadata.TruncateBefore] = recreateFromEventNumber;
+                using (var memoryStream = new MemoryStream())
+                {
+                    using (var jsonWriter = new JsonTextWriter(new StreamWriter(memoryStream)))
+                    {
+                        jobj.WriteTo(jsonWriter);
+                    }
+                    modifiedMeta = memoryStream.ToArray();
+                    return true;
+                }
+            }
+            catch (Exception)
+            {
+                modifiedMeta = null;
+                return false;
+            }
+
+        }
+
         void IHandle<StorageMessage.WriteDelete>.Handle(StorageMessage.WriteDelete message)
         {
             Interlocked.Decrement(ref FlushMessagesInQueue);
@@ -329,18 +391,36 @@ namespace EventStore.Core.Services.Storage
 
                 var eventId = Guid.NewGuid();
                 
-                var commitCheck = IndexWriter.CheckCommit(message.EventStreamId, message.ExpectedVersion, new[]{eventId});
+                var commitCheck = _indexWriter.CheckCommit(message.EventStreamId, message.ExpectedVersion, new[]{eventId});
                 if (commitCheck.Decision != CommitDecision.Ok)
                 {
                     ActOnCommitCheckFailure(message.Envelope, message.CorrelationId, commitCheck);
                     return;
                 }
-                // when IsCommitted ExpectedVersion is actually real EventNumber
-                const int expectedVersion = EventNumber.DeletedStream - 1;
-                var record = LogRecord.DeleteTombstone(Writer.Checkpoint.ReadNonFlushed(), message.CorrelationId,
-                                                       eventId, message.EventStreamId, expectedVersion, PrepareFlags.IsCommitted); 
-                var res = WritePrepareWithRetry(record);
-                IndexWriter.PreCommit(new[] {res.Prepare});
+
+                if (message.HardDelete)
+                {
+                    // HARD DELETE
+                    const int expectedVersion = EventNumber.DeletedStream - 1;
+                    var record = LogRecord.DeleteTombstone(Writer.Checkpoint.ReadNonFlushed(), message.CorrelationId,
+                                                           eventId, message.EventStreamId, expectedVersion, PrepareFlags.IsCommitted);
+                    var res = WritePrepareWithRetry(record);
+                    _indexWriter.PreCommit(new[] { res.Prepare });
+                }
+                else 
+                {
+                    // SOFT DELETE
+                    var metastreamId = SystemStreams.MetastreamOf(message.EventStreamId);
+                    var expectedVersion = _indexWriter.GetStreamLastEventNumber(metastreamId);
+                    var logPosition = Writer.Checkpoint.ReadNonFlushed();
+                    const PrepareFlags flags = PrepareFlags.SingleWrite | PrepareFlags.IsCommitted | PrepareFlags.IsJson;
+                    var data = new StreamMetadata(truncateBefore: EventNumber.DeletedStream).ToJsonBytes();
+                    var res = WritePrepareWithRetry(
+                        LogRecord.Prepare(logPosition, message.CorrelationId, eventId, logPosition, 0,
+                                          metastreamId, expectedVersion, flags, SystemEventTypes.StreamMetadata,
+                                          data, null));
+                    _indexWriter.PreCommit(new[] { res.Prepare });
+                }
             }
             catch (Exception exc)
             {
@@ -368,7 +448,7 @@ namespace EventStore.Core.Services.Storage
                 var res = WritePrepareWithRetry(record);
 
                 // we update cache to avoid non-cached look-up on next TransactionWrite
-                IndexWriter.UpdateTransactionInfo(res.WrittenPos, res.WrittenPos, new TransactionInfo(-1, message.EventStreamId)); 
+                _indexWriter.UpdateTransactionInfo(res.WrittenPos, res.WrittenPos, new TransactionInfo(-1, message.EventStreamId)); 
             }
             catch (Exception exc)
             {
@@ -387,7 +467,7 @@ namespace EventStore.Core.Services.Storage
             try
             {
                 var logPosition = Writer.Checkpoint.ReadNonFlushed();
-                var transactionInfo = IndexWriter.GetTransactionInfo(Writer.Checkpoint.Read(), message.TransactionId);
+                var transactionInfo = _indexWriter.GetTransactionInfo(Writer.Checkpoint.Read(), message.TransactionId);
                 if (!CheckTransactionInfo(message.TransactionId, transactionInfo))
                     return;
 
@@ -413,7 +493,7 @@ namespace EventStore.Core.Services.Storage
                     }
                     var info = new TransactionInfo(transactionInfo.TransactionOffset + message.Events.Length,
                                                    transactionInfo.EventStreamId);
-                    IndexWriter.UpdateTransactionInfo(message.TransactionId, lastLogPosition, info);
+                    _indexWriter.UpdateTransactionInfo(message.TransactionId, lastLogPosition, info);
                 }
             }
             catch (Exception exc)
@@ -435,7 +515,7 @@ namespace EventStore.Core.Services.Storage
                 if (message.LiveUntil < DateTime.UtcNow)
                     return;
 
-                var transactionInfo = IndexWriter.GetTransactionInfo(Writer.Checkpoint.Read(), message.TransactionId);
+                var transactionInfo = _indexWriter.GetTransactionInfo(Writer.Checkpoint.Read(), message.TransactionId);
                 if (!CheckTransactionInfo(message.TransactionId, transactionInfo))
                     return;
 
@@ -477,16 +557,28 @@ namespace EventStore.Core.Services.Storage
             try
             {
                 var commitPos = Writer.Checkpoint.ReadNonFlushed();
-                var result = IndexWriter.CheckCommitStartingAt(message.TransactionPosition, commitPos);
-                if (result.Decision == CommitDecision.Ok)
+                var commitCheck = _indexWriter.CheckCommitStartingAt(message.TransactionPosition, commitPos);
+                if (commitCheck.Decision != CommitDecision.Ok)
                 {
-                    var commit = WriteCommitWithRetry(LogRecord.Commit(commitPos,
-                                                                       message.CorrelationId,
-                                                                       message.TransactionPosition,
-                                                                       result.CurrentVersion + 1));
-                    IndexWriter.PreCommit(commit);
+                    ActOnCommitCheckFailure(message.Envelope, message.CorrelationId, commitCheck);
+                    return;
                 }
-                else ActOnCommitCheckFailure(message.Envelope, message.CorrelationId, result);
+
+
+                var commit = WriteCommitWithRetry(LogRecord.Commit(commitPos,
+                                                                   message.CorrelationId,
+                                                                   message.TransactionPosition,
+                                                                   commitCheck.CurrentVersion + 1));
+
+                bool softUndeleteMetastream = SystemStreams.IsMetastream(commitCheck.EventStreamId)
+                                              && _indexWriter.IsSoftDeleted(SystemStreams.OriginalStreamOf(commitCheck.EventStreamId));
+               
+                _indexWriter.PreCommit(commit);
+
+                if (commitCheck.IsSoftDeleted)
+                    SoftUndeleteStream(commitCheck.EventStreamId, commitCheck.CurrentVersion + 1);
+                if (softUndeleteMetastream)
+                    SoftUndeleteMetastream(commitCheck.EventStreamId);
             }
             catch (Exception exc)
             {
@@ -623,8 +715,8 @@ namespace EventStore.Core.Services.Storage
 
         private void PurgeNotProcessedInfo()
         {
-            IndexWriter.PurgeNotProcessedCommitsTill(Db.Config.ChaserCheckpoint.Read());
-            IndexWriter.PurgeNotProcessedTransactions(Db.Config.WriterCheckpoint.Read());
+            _indexWriter.PurgeNotProcessedCommitsTill(Db.Config.ChaserCheckpoint.Read());
+            _indexWriter.PurgeNotProcessedTransactions(Db.Config.WriterCheckpoint.Read());
         }
 
         private struct WriteResult
